Create panel store
